// applets/sonarr/api/model/episodeFileResource.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { CustomFormatResource } from "./customFormatResource";
import type { Language } from "./language";
import type { MediaInfoResource } from "./mediaInfoResource";
import type { QualityModel } from "./qualityModel";

export interface EpisodeFileResource {
  /** @nullable */
  customFormats?: CustomFormatResource[] | null;
  dateAdded?: string;
  id?: number;
  /** @nullable */
  languages?: Language[] | null;
  mediaInfo?: MediaInfoResource;
  /** @nullable */
  path?: string | null;
  quality?: QualityModel;
  qualityCutoffNotMet?: boolean;
  /** @nullable */
  relativePath?: string | null;
  /** @nullable */
  releaseGroup?: string | null;
  /** @nullable */
  sceneName?: string | null;
  seasonNumber?: number;
  seriesId?: number;
  size?: number;
}