// applets/sonarr/api/model/episodeResource.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { EpisodeFileResource } from "./episodeFileResource";
import type { SeriesResource } from "./seriesResource";
import type { MediaCover } from "./mediaCover";

export interface EpisodeResource {
  id?: number;
  seriesId?: number;
  tvdbId?: number;
  episodeFileId?: number;
  seasonNumber?: number;
  episodeNumber?: number;
  title?: string | null;
  airDate?: string | null;
  airDateUtc?: string | null;
  runtime?: number;
  overview?: string | null;
  episodeFile?: EpisodeFileResource;
  hasFile?: boolean;
  monitored?: boolean;
  absoluteEpisodeNumber?: number | null;
  sceneAbsoluteEpisodeNumber?: number | null;
  sceneEpisodeNumber?: number | null;
  sceneSeasonNumber?: number | null;
  unverifiedSceneNumbering?: boolean;
  endTime?: string | null;
  grabDate?: string | null;
  seriesTitle?: string | null;
  series?: SeriesResource;
  images?: MediaCover[] | null;
  grabbed?: boolean;
}